Build sensor list treeview from sensor names
Splitting sensor names on dots to build up the tree.  This uses
the filtered list of sensors.

Tree is initially fully collapsed, and only the first level of nodes
are available to the ui tree.  A form of lazy loading.  This speeds up
the initial display.  As each subtree is expanded, only one level of
nodes in that branch are "loaded".

Added test buttons to collapse all and expand all nodes.  Still WIP.