Filter by most used user product
The users have preference of the user product(which should be default)
that is not the first on our list of user products per band. We have
implemented a hack to make the preference the default.

There is also a trade-off between site and dev environments in this,
site does not have the `_test_...` type of user product while our
dev environments do. The hack will work on site and devs will have to
do the product selection in the case of `x-band` and `s-band` which only
have the `_test_...` user product. We should be able to live with this